+++ setup.sh
@@ -1,2 +1,3 @@
+docker system prune
 docker build -t praise-setup -f ./packages/setup/Dockerfile .
 docker run -it -v $(pwd):/usr/praise praise-setup