+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Quantifiers displayed as not being done when they in fact are done #349
+- UI issue: checkbox unchecked when praise score changes #354
 
 ## [0.5.0] - 2022-05-02