+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Fixed
+
+## [0.6.0] - 2022-05-09
+
+### Added
+
 - Add giver and receiver id to exported praise csv #364
 - Increased precision in calculation of duplicate praise score, from 0 to 2 decimals #342